create AnalogOutClass as separate module
